Update campaign add & edit screen
