+++ includes/Admin/views/campaigns/add.php
@@ -10,8 +10,8 @@ defined( 'ABSPATH' ) || exit;
 
 ?>
 <h1><?php esc_html_e( 'Add New Campaign', 'wc-donation-manager' ); ?></h1>
-<p><?php esc_html_e( 'You can create a new campaign here. This form will create a campaign/donation product. It will be available on woocommerce products page as well for latter edit or updates.', 'wc-donation-manager' ); ?></p>
-<form method="POST" action="<?php echo esc_url( admin_url( 'admin-post.php' ) ); ?>">
+<p><?php esc_html_e( 'You can create a new campaign here. This form will create a campaign.', 'wc-donation-manager' ); ?></p>
+<form class="campaign" method="POST" action="<?php echo esc_url( admin_url( 'admin-post.php' ) ); ?>">
 	<table class="form-table">
 		<tbody>
 		<tr valign="top">
@@ -30,7 +30,7 @@ defined( 'ABSPATH' ) || exit;
 				<label for="cause"><?php esc_html_e( 'Cause', 'wc-donation-manager' ); ?></label>
 			</th>
 			<td>
-				<textarea name="cause" id="cause" rows="6" placeholder="<?php esc_html_e( 'Enter the cause of the campaign...', 'wc-donation-manager' ); ?>"></textarea>
+				<textarea name="cause" id="cause" class="regular-text" rows="6" placeholder="<?php esc_html_e( 'Enter the cause of the campaign...', 'wc-donation-manager' ); ?>"></textarea>
 				<p class="description">
 					<?php esc_html_e( 'Enter the cause of the campaign.', 'wc-donation-manager' ); ?>
 				</p>
@@ -38,7 +38,7 @@ defined( 'ABSPATH' ) || exit;
 		</tr>
 		<tr valign="top">
 			<th scope="row">
-				<label for="goal_amount"><?php esc_html_e( 'Goal amount', 'wc-donation-manager' ); ?></label>
+				<label for="goal_amount"><?php esc_html_e( 'Goal amount *', 'wc-donation-manager' ); ?></label>
 			</th>
 			<td>
 				<input type="number" min="0" step="any" name="goal_amount" id="goal_amount" class="regular-text" value="" required/>
@@ -49,7 +49,7 @@ defined( 'ABSPATH' ) || exit;
 		</tr>
 		<tr valign="top">
 			<th scope="row">
-				<label for="end_date"><?php esc_html_e( 'End date', 'wc-donation-manager' ); ?></label>
+				<label for="end_date"><?php esc_html_e( 'End date *', 'wc-donation-manager' ); ?></label>
 			</th>
 			<td>
 				<input type="date" name="end_date" id="end_date" class="regular-text" value="" required/>
@@ -63,10 +63,10 @@ defined( 'ABSPATH' ) || exit;
 				<label for="status"><?php esc_html_e( 'Status', 'wc-donation-manager' ); ?></label>
 			</th>
 			<td>
-				<select id="status" name="status" style="width: 300px" required>
-					<option value="Publish">Publish</option>
-					<option value="Pending">Pending</option>
-					<option value="Draft">Draft</option>
+				<select id="status" name="status" class="regular-text" required>
+					<option value="publish">Publish</option>
+					<option value="pending">Pending</option>
+					<option value="draft">Draft</option>
 				</select>
 				<p class="description">
 					<?php esc_html_e( 'Select the campaign status.', 'wc-donation-manager' ); ?>
