+++ languages/wc-donation-manager.pot
@@ -4,7 +4,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: WC Donation Manager 1.0.1\n"
 "Report-Msgid-Bugs-To: https://pluginever.com/plugins/wc-donation-manager/\n"
-"POT-Creation-Date: 2024-07-27 10:19:47+00:00\n"
+"POT-Creation-Date: 2024-07-28 03:58:04+00:00\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
@@ -24,19 +24,23 @@ msgstr ""
 "X-Textdomain-Support: yes\n"
 "X-Generator: grunt-wp-i18n 1.0.3\n"
 
-#: includes/Admin/Actions.php:51 includes/Admin/Actions.php:76
+#: includes/Admin/Actions.php:82
 msgid "Campaign saved successfully."
 msgstr ""
 
-#: includes/Admin/Actions.php:133
+#: includes/Admin/Actions.php:123
+msgid "Campaign updated successfully."
+msgstr ""
+
+#: includes/Admin/Actions.php:150
 msgid "Suggested amounts"
 msgstr ""
 
-#: includes/Admin/Actions.php:157 includes/Admin/Actions.php:177
+#: includes/Admin/Actions.php:174 includes/Admin/Actions.php:194
 msgid "Donor saved successfully."
 msgstr ""
 
-#: includes/Admin/Actions.php:194
+#: includes/Admin/Actions.php:211
 msgid "No, search term provided."
 msgstr ""
 
@@ -81,7 +85,6 @@ msgid "Goal"
 msgstr ""
 
 #: includes/Admin/ListTables/CampaignsListTable.php:96
-#: includes/Admin/views/campaigns/add.php:52
 msgid "End date"
 msgstr ""
 
@@ -464,10 +467,7 @@ msgid "Add New Campaign"
 msgstr ""
 
 #: includes/Admin/views/campaigns/add.php:13
-msgid ""
-"You can create a new campaign here. This form will create a "
-"campaign/donation product. It will be available on woocommerce products "
-"page as well for latter edit or updates."
+msgid "You can create a new campaign here. This form will create a campaign."
 msgstr ""
 
 #: includes/Admin/views/campaigns/add.php:19
@@ -490,11 +490,17 @@ msgid "Enter the cause of the campaign."
 msgstr ""
 
 #: includes/Admin/views/campaigns/add.php:41
-msgid "Goal amount"
+msgid "Goal amount *"
 msgstr ""
 
 #: includes/Admin/views/campaigns/add.php:46
 msgid "Enter the goal amount of the campaign."
+msgstr ""
+
+#: includes/Admin/views/campaigns/add.php:52
+#: includes/Admin/views/campaigns/edit.php:67
+#: includes/Admin/views/campaigns/view.php:61
+msgid "End date *"
 msgstr ""
 
 #: includes/Admin/views/campaigns/add.php:57
@@ -555,11 +561,6 @@ msgstr ""
 msgid "Edit or update the goal amount of the campaign."
 msgstr ""
 
-#: includes/Admin/views/campaigns/edit.php:67
-#: includes/Admin/views/campaigns/view.php:61
-msgid "End date *"
-msgstr ""
-
 #: includes/Admin/views/campaigns/edit.php:72
 msgid "Edit or update the end date of the campaign."
 msgstr ""
@@ -581,12 +582,6 @@ msgstr ""
 msgid "%1$s%2$.2f goal"
 msgstr ""
 
-#: includes/Admin/views/campaigns/edit.php:89
-#: includes/Admin/views/campaigns/view.php:83 includes/Frontend/Product.php:94
-#. translators: 1: Raised amount
-msgid "%1$s%%"
-msgstr ""
-
 #: includes/Admin/views/campaigns/edit.php:94
 #: includes/Admin/views/campaigns/view.php:88
 msgid "Actions"
@@ -597,7 +592,7 @@ msgid "Update the campaign status."
 msgstr ""
 
 #: includes/Admin/views/campaigns/edit.php:114
-msgid "Save Campaign"
+msgid "Save Changes"
 msgstr ""
 
 #: includes/Admin/views/campaigns/edit.php:129
@@ -624,6 +619,11 @@ msgstr ""
 
 #: includes/Admin/views/campaigns/view.php:66
 msgid "The end date of the campaign."
+msgstr ""
+
+#: includes/Admin/views/campaigns/view.php:83
+#. translators: 1: Raised amount
+msgid "%1$s%%"
 msgstr ""
 
 #: includes/Admin/views/campaigns/view.php:92
@@ -714,11 +714,6 @@ msgstr ""
 msgid ""
 "Just to let you know — we've received your donation order #{order_number}, "
 "and it is now being processed:"
-msgstr ""
-
-#: includes/Frontend/Product.php:122
-#. translators: 1: WC currency symbol
-msgid "%1$s"
 msgstr ""
 
 #: includes/Frontend/Product.php:144
