+++ src/css/admin-common.scss
@@ -31,6 +31,11 @@
 			border-bottom: none;
 		}
 	}
+	.form-table {
+		.regular-text {
+			width: 25em!important;
+		}
+	}
 }
 .campaign-progress {
 	.progress-label {
